Theme config moved to a seperate file
