+++ src/theme/ThemeConfig.tsx
@@ -1,0 +1,12 @@
+import { blue, pink, orange, red } from '@material-ui/core/colors';
+import { IThemeConfig } from './IThemeConfig';
+
+export const ThemeConfig: IThemeConfig = {
+  isGradientEnabled: true,
+  primaryColor: blue,
+  secondaryColor: pink,
+  warningColor: orange,
+  dangerColor: red,
+  direction: 'ltr',
+  drawerWidth: 240,
+};
